Updater: fix local file installation
Local file installations require that the file
is already on the device. Moreover, adding a local
file also verifies it during the process.
Hence, the added file is in a dormant non-active state
and must be treated as such.

Secondly, a successful verification is hardly an error.

Thirdly, use filetime instead of current time when
setting update timestamp, so we can prevent older builds
to be manually flashed unless the allow-downgrade system
property is set.

Finally, local file versions are empty strings and
confuse our Utils canInstall function, which expects
a fully formed version string, causing it to return a nay
for canInstall. Check update's persistent
status instead of version in local updates.

Change-Id: I5d64534c4cd8fda48121cf6ae4c422a706187d42